moved ensemble models to core
