+++ core/src/autogluon/core/models/__init__.py
@@ -1,1 +1,5 @@
 from .abstract.abstract_model import AbstractModel
+from .greedy_ensemble.greedy_weighted_ensemble_model import GreedyWeightedEnsembleModel
+from .ensemble.bagged_ensemble_model import BaggedEnsembleModel
+from .ensemble.stacker_ensemble_model import StackerEnsembleModel
+from .ensemble.weighted_ensemble_model import WeightedEnsembleModel
