+++ core/src/autogluon/core/models/greedy_ensemble/ensemble_selection.py
@@ -1,10 +1,12 @@
-import logging, time
-import numpy as np
+import logging
+import time
 from collections import Counter
 
-from autogluon.core.constants import PROBLEM_TYPES
-from autogluon.core.metrics import calculate_score, _ProbaScorer, _ThresholdScorer
-from autogluon.core.utils import get_pred_from_proba
+import numpy as np
+
+from ...constants import PROBLEM_TYPES
+from ...metrics import calculate_score, _ProbaScorer, _ThresholdScorer
+from ...utils import get_pred_from_proba
 
 logger = logging.getLogger(__name__)
