+++ tabular/src/autogluon/tabular/models/__init__.py
@@ -9,11 +9,5 @@ from .knn.knn_model import KNNModel
 from .lr.lr_model import LinearModel
 from .tabular_nn.tabular_nn_model import TabularNeuralNetModel
 from .fastainn.tabular_nn_fastai import NNFastAiTabularModel
-
-from .ensemble.bagged_ensemble_model import BaggedEnsembleModel
-from .ensemble.stacker_ensemble_model import StackerEnsembleModel
-from .ensemble.weighted_ensemble_model import WeightedEnsembleModel
-from .ensemble.greedy_weighted_ensemble_model import GreedyWeightedEnsembleModel
-
 from .fasttext.fasttext_model import FastTextModel
 from .text_prediction.text_prediction_v1_model import TextPredictionV1Model
