+++ tabular/src/autogluon/tabular/trainer/model_presets/presets.py
@@ -1,27 +1,15 @@
 import copy
 import inspect
 import logging
-import warnings
 from collections import defaultdict
 
 from autogluon.core.metrics import soft_log_loss, mean_squared_error
-from autogluon.core.constants import AG_ARGS, AG_ARGS_FIT, AG_ARGS_ENSEMBLE, BINARY, MULTICLASS,\
-    REGRESSION, SOFTCLASS
-from autogluon.core.models import AbstractModel
-from ...models.fastainn.tabular_nn_fastai import NNFastAiTabularModel
-from ...models.lgb.lgb_model import LGBModel
-from ...models.lr.lr_model import LinearModel
-from ...models.tabular_nn.tabular_nn_model import TabularNeuralNetModel
-from ...models.rf.rf_model import RFModel
-from ...models.knn.knn_model import KNNModel
-from ...models.catboost.catboost_model import CatBoostModel
-from ...models.xgboost.xgboost_model import XGBoostModel
-from ...models.xt.xt_model import XTModel
+from autogluon.core.constants import AG_ARGS, AG_ARGS_FIT, AG_ARGS_ENSEMBLE, BINARY, MULTICLASS, REGRESSION, SOFTCLASS
+from autogluon.core.models import AbstractModel, GreedyWeightedEnsembleModel, StackerEnsembleModel
+
+from ...models import LGBModel, CatBoostModel, XGBoostModel, RFModel, XTModel, KNNModel, LinearModel,\
+    TabularNeuralNetModel, NNFastAiTabularModel, FastTextModel, TextPredictionV1Model
 from ...models.tab_transformer.tab_transformer_model import TabTransformerModel
-from ...models.fasttext.fasttext_model import FastTextModel
-from ...models.text_prediction.text_prediction_v1_model import TextPredictionV1Model
-from ...models.ensemble.stacker_ensemble_model import StackerEnsembleModel
-from ...models.ensemble.greedy_weighted_ensemble_model import GreedyWeightedEnsembleModel
 
 logger = logging.getLogger(__name__)
